# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Rapide API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/cloudflare/dns/UpdateDnsRecord": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cloudflare"
                ],
                "summary": "更新Dns记录",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cloudflare.DnsUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/cloudflare/dns/createDnsRecord": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cloudflare"
                ],
                "summary": "创建Dns记录",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cloudflare.DnsCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/cloudflare/dns/deleteDnsRecord": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cloudflare"
                ],
                "summary": "删除Dns记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "zone_id",
                        "name": "zone_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dns_id",
                        "name": "dns_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/cloudflare/dns/getDnsList": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cloudflare"
                ],
                "summary": "获取dns列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "zone_id",
                        "name": "zone_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/cloudflare/dns/getDnsRecordDetail": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cloudflare"
                ],
                "summary": "获取Dns记录详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "zone_id",
                        "name": "zone_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dns_id",
                        "name": "dns_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/dept/addDept": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "新增部门",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sys.DeptAddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/dept/deleteDept": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sys.DeptDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/dept/getDept": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理"
                ],
                "summary": "获取部门列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cloudflare.DnsCreateRequest": {
            "type": "object",
            "required": [
                "content",
                "name",
                "type",
                "zone_id"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "maxLength": 100
                },
                "content": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "proxied": {
                    "type": "boolean"
                },
                "ttl": {
                    "type": "integer",
                    "maximum": 3600,
                    "minimum": 0
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "A",
                        "AAAA",
                        "CNAME",
                        "MX",
                        "NS",
                        "TXT"
                    ]
                },
                "zone_id": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 0
                }
            }
        },
        "cloudflare.DnsUpdateRequest": {
            "type": "object",
            "required": [
                "content",
                "dns_id",
                "name",
                "type",
                "zone_id"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "maxLength": 100
                },
                "content": {
                    "type": "string"
                },
                "dns_id": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "proxied": {
                    "type": "boolean"
                },
                "ttl": {
                    "type": "integer",
                    "maximum": 3600,
                    "minimum": 0
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "A",
                        "AAAA",
                        "CNAME",
                        "MX",
                        "NS",
                        "TXT"
                    ]
                },
                "zone_id": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 0
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "sys.DeptAddRequest": {
            "type": "object",
            "required": [
                "fullname",
                "num",
                "pid",
                "pids",
                "tips"
            ],
            "properties": {
                "fullname": {
                    "type": "string"
                },
                "num": {
                    "type": "integer"
                },
                "pid": {
                    "type": "integer"
                },
                "pids": {
                    "type": "string"
                },
                "tips": {
                    "type": "string"
                }
            }
        },
        "sys.DeptDeleteRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}